requirements.txt
ccxt
pandas
numpy
python-telegram-bot==20.3
